+++ src/main.js
@@ -9,7 +9,7 @@ import StatsView from './view/stats.js';
 import FilmListEmptyView from './view/film-list-empry.js';
 import {createFilmCard} from './mock/film-card.js';
 import {generateFilter} from './mock/filter.js';
-import {render, RenderPosition} from './utils.js';
+import {render, RenderPosition, remove} from './utils/render.js';
 
 const CARDS_COUNT = 17;
 const FILMS_COUNT_PER_STEP = 5;
@@ -24,7 +24,7 @@ const bodyElement = document.querySelector(`body`);
 const renderFilm = (filmListElement, film) => {
   const filmComponent = new FilmCardView(film);
 
-  render(filmListElement, filmComponent.getElement(), RenderPosition.BEFOREEND);
+  render(filmListElement, filmComponent, RenderPosition.BEFOREEND);
 
   const filmElementMatches = `.film-card__title, .film-card__poster, .film-card__comments`;
   const popup = new PopupView(film);
@@ -61,14 +61,14 @@ const renderFilm = (filmListElement, film) => {
   filmComponent.setClickOpenHandler(popupOpenHandler);
 };
 
-render(siteHeader, new UserProfileView().getElement(), RenderPosition.BEFOREEND);
-render(siteMain, new SiteNavigationView(filters).getElement(), RenderPosition.BEFOREEND);
+render(siteHeader, new UserProfileView(), RenderPosition.BEFOREEND);
+render(siteMain, new SiteNavigationView(filters), RenderPosition.BEFOREEND);
 
 if (films.length === 0) {
-  render(siteMain, new FilmListEmptyView().getElement(), RenderPosition.BEFOREEND);
+  render(siteMain, new FilmListEmptyView(), RenderPosition.BEFOREEND);
 } else {
-  render(siteMain, new SiteSortingView().getElement(), RenderPosition.BEFOREEND);
-  render(siteMain, new FilmListView().getElement(), RenderPosition.BEFOREEND);
+  render(siteMain, new SiteSortingView(), RenderPosition.BEFOREEND);
+  render(siteMain, new FilmListView(), RenderPosition.BEFOREEND);
 
   const filmsList = siteMain.querySelector(`.films-list`);
   const filmsListContainer = filmsList.querySelector(`.films-list__container`);
@@ -77,7 +77,7 @@ if (films.length === 0) {
     let renderedFilmsCount = FILMS_COUNT_PER_STEP;
     const showMoreButtonComponent = new ShowMoreButtonView();
 
-    render(filmsList, showMoreButtonComponent.getElement(), RenderPosition.BEFOREEND);
+    render(filmsList, showMoreButtonComponent, RenderPosition.BEFOREEND);
 
     showMoreButtonComponent.setClickHandler(() => {
       films
@@ -87,8 +87,7 @@ if (films.length === 0) {
       renderedFilmsCount += FILMS_COUNT_PER_STEP;
 
       if (renderedFilmsCount >= films.length) {
-        showMoreButtonComponent.getElement().remove();
-        showMoreButtonComponent.removeElement();
+        remove(showMoreButtonComponent);
       }
     });
   }
@@ -98,7 +97,7 @@ if (films.length === 0) {
   }
 }
 
-render(siteMain, new StatsView().getElement(), RenderPosition.BEFOREEND);
+render(siteMain, new StatsView(), RenderPosition.BEFOREEND);
 
 const createFooterStatistics = (filmsCount) => {
   return `${filmsCount} movies inside`;
