+++ src/utils/common.js
@@ -1,0 +1,17 @@
+
+import dayjs from 'dayjs';
+
+export const getRandomInteger = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+export const getRandomItem = (arr) => arr[Math.floor(Math.random() * arr.length)];
+
+export const generateDate = () => {
+  const maxDaysGap = 7;
+  const daysGap = getRandomInteger(-maxDaysGap, maxDaysGap);
+
+  return dayjs().add(daysGap, `day`).format(`DD MMMM YYYY`);
+};
